refactor: move the register and unregister functions into the "setup.ts" file
